+++ src/errors/utils/deserializeError.ts
@@ -1,0 +1,18 @@
+import { ERRORS } from '..';
+import { ErrorJson } from './ErrorJson';
+
+/**
+ * Deserializes the error object
+ *
+ * @public exported from `@promptbook/utils`
+ */
+
+export function deserializeError(error: ErrorJson): Error {
+    if (error.name === 'Error') {
+        return new Error(error.message);
+    }
+
+    const CustomError = ERRORS[error.name as keyof typeof ERRORS];
+
+    return new CustomError(error.message);
+}