+++ src/types/PipelineJson/KnowledgePieceJson.ts
@@ -1,17 +1,19 @@
 import type { EmbeddingVector } from '../../execution/EmbeddingVector';
 import type { string_keyword } from '../../utils/normalization/IKeywords';
-import type { number_id } from '../typeAliases';
-import type { number_linecol_number } from '../typeAliases';
-import type { string_markdown } from '../typeAliases';
-import type { string_markdown_text } from '../typeAliases';
-import type { string_model_name } from '../typeAliases';
-import type { string_name } from '../typeAliases';
+import type {
+    number_id,
+    number_linecol_number,
+    string_markdown,
+    string_markdown_text,
+    string_model_name,
+    string_name,
+} from '../typeAliases';
 
 /**
  * Defines one piece of knowledge in the pipeline
  *
  * Note: Knowledge piece is by definition prepared
- * 
+ *
  * Note: [🚉] This is fully serializable as JSON
  *
  * @see https://github.com/webgptorg/promptbook/discussions/41