+++ src/types/PipelineJson/KnowledgeSourceJson.ts
@@ -1,11 +1,9 @@
-import type { number_id } from '../typeAliases';
-import type { string_knowledge_source_content } from '../typeAliases';
-import type { string_name } from '../typeAliases';
+import type { number_id, string_knowledge_source_content, string_name } from '../typeAliases';
 
 /**
  * Defines one source of knowledge in the pipeline
  * For example, a source of information, a fact, a quote, a definition, website, etc.
- * 
+ *
  * Note: [🚉] This is fully serializable as JSON
  *
  * @see https://github.com/webgptorg/promptbook/discussions/41
@@ -24,7 +22,7 @@ export type KnowledgeSourceJson = {
 
 /**
  * Defines one source of knowledge in the pipeline after it has been prepared
- * 
+ *
  * Note: [🚉] This is fully serializable as JSON
  *
  * @see https://github.com/webgptorg/promptbook/discussions/41