+++ src/types/PipelineJson/PersonaJson.ts
@@ -1,11 +1,9 @@
 import type { ChatModelRequirements } from '../ModelRequirements';
-import type { number_id } from '../typeAliases';
-import type { string_name } from '../typeAliases';
-import type { string_persona_description } from '../typeAliases';
+import type { number_id, string_name, string_persona_description } from '../typeAliases';
 
 /**
  * Defines a persona in the pipeline
- * 
+ *
  * Note: [🚉] This is fully serializable as JSON
  *
  * @see https://github.com/webgptorg/promptbook/discussions/22
@@ -26,7 +24,7 @@ export type PersonaJson = {
 
 /**
  * Defines a persona in the pipeline after it has been prepared
- * 
+ *
  * Note: [🚉] This is fully serializable as JSON
  *
  * @see https://github.com/webgptorg/promptbook/discussions/22